+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added new validator `expressionIsWellFormatted` to check if rules are well formatted as `promtool promql format` would do.
 
 ## [v2.8.1] - 2024-02-29
 - Fixed param validation of the `hasAllowedEvaluationInterval` validator, if the `maximum` was not set.